+++ packages/react-native-sdk/src/hooks/index.ts
@@ -1,5 +1,5 @@
 export { useConnect } from "./useConnect";
 export { useDisconnect } from "./useDisconnect";
 export { useAccounts } from "./useAccounts";
-export { useSignMessage } from "./useSignMessage";
-export { useSignAndSendTransaction } from "./useSignAndSendTransaction";
+export { useSolana } from "./useSolana";
+export { useEthereum } from "./useEthereum";