+++ packages/react-native-sdk/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @phantom/react-native-sdk
 
+## 1.0.0-beta.0
+
+### Major Changes
+
+- New SDKs chain specific interface, prepare for initial beta release
+
+### Patch Changes
+
+- ef68eaf: Key rotation
+- Updated dependencies [ef68eaf]
+- Updated dependencies
+  - @phantom/embedded-provider-core@1.0.0-beta.0
+  - @phantom/sdk-types@1.0.0-beta.0
+  - @phantom/client@1.0.0-beta.0
+  - @phantom/constants@1.0.0-beta.0
+  - @phantom/parsers@1.0.0-beta.0
+  - @phantom/chains@1.0.0-beta.0
+  - @phantom/api-key-stamper@1.0.0-beta.0
+  - @phantom/base64url@1.0.0-beta.0
+  - @phantom/crypto@1.0.0-beta.0
+
+## 0.1.8
+
+### Patch Changes
+
+- b20bc21: Support derivation index
+- Updated dependencies [b20bc21]
+  - @phantom/embedded-provider-core@0.1.10
+  - @phantom/client@0.1.11
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @phantom/embedded-provider-core@0.1.9
+  - @phantom/sdk-types@0.1.5
+  - @phantom/client@0.1.10
+
 ## 0.1.6
 
 ### Patch Changes